+++ examples/demos/basic.js
@@ -5,12 +5,11 @@ import events from '../events';
 let Basic = React.createClass({
   render(){
     return (
-      <div>
-        <BigCalendar
-          events={events}
-          defaultDate={new Date(2015, 3, 1)}
-        />
-      </div>
+      <BigCalendar
+        {...this.props}
+        events={events}
+        defaultDate={new Date(2015, 3, 1)}
+      />
     )
   }
 })