+++ src/TimeGrid.jsx
@@ -1,15 +1,14 @@
-import React from 'react';
+import React, { Component } from 'react';
 import cn from 'classnames';
 import { findDOMNode } from 'react-dom';
 import dates from './utils/dates';
 import localizer from './localizer'
 
-import DaySlot from './DaySlot';
+import DayColumn from './DayColumn';
 import EventRow from './EventRow';
-import TimeGutter from './TimeGutter';
+import TimeColumn from './TimeColumn';
 import BackgroundCells from './BackgroundCells';
 
-import classes from 'dom-helpers/class';
 import getWidth from 'dom-helpers/query/width';
 import scrollbarSize from 'dom-helpers/util/scrollbarSize';
 import message from './utils/messages';
@@ -20,49 +19,73 @@ import { notify } from './utils/helpers';
 import { navigate } from './utils/constants';
 import { accessor as get } from './utils/accessors';
 
-import { inRange, eventSegments, eventLevels, sortEvents, segStyle } from './utils/eventLevels';
+import {
+  inRange, eventSegments, endOfRange
+  , eventLevels, sortEvents, segStyle } from './utils/eventLevels';
 
 const MIN_ROWS = 2;
 
-let TimeGrid = React.createClass({
 
-  propTypes: {
-    ...DaySlot.propTypes,
-    ...TimeGutter.propTypes,
+export default class TimeGrid extends Component {
+
+  static propTypes = {
+    ...DayColumn.propTypes,
+    ...TimeColumn.propTypes,
 
     step: React.PropTypes.number,
     min: React.PropTypes.instanceOf(Date),
     max: React.PropTypes.instanceOf(Date),
     dayFormat: dateFormat,
     rtl: React.PropTypes.bool
-  },
+  }
 
-  getDefaultProps(){
-    return {
-      step: 30,
-      min: dates.startOf(new Date(), 'day'),
-      max: dates.endOf(new Date(), 'day')
-    }
-  },
+  static defaultProps = {
+    ...DayColumn.defaultProps,
+    ...TimeColumn.defaultProps,
+
+    step: 30,
+    min: dates.startOf(new Date(), 'day'),
+    max: dates.endOf(new Date(), 'day'),
+    /* these 2 are needed to satisfy requirements from TimeColumn required props
+     * There is a strange bug in React, using ...TimeColumn.defaultProps causes weird crashes
+     */
+    type: 'gutter',
+    now: new Date()
+  }
+
+  constructor(props) {
+    super(props)
+    this.state = { gutterWidth: undefined, isOverflowing: null };
+    this._selectEvent = this._selectEvent.bind(this)
+    this._headerClick = this._headerClick.bind(this)
+  }
 
   componentWillMount() {
     this._gutters = [];
-  },
+  }
 
   componentDidMount() {
-    this._adjustGutter()
-  },
+    this.checkOverflow();
+
+    if (this.props.width == null) {
+      this.measureGutter()
+    }
+  }
 
   componentDidUpdate() {
-    this._adjustGutter()
-  },
+    if (this.props.width == null && !this.state.gutterWidth) {
+      this.measureGutter()
+    }
+
+    //this.checkOverflow()
+  }
 
   render() {
     let {
-        events, start, end, messages
+        events, start, end, width
       , startAccessor, endAccessor, allDayAccessor } = this.props;
 
-    let addGutterRef = i => ref => this._gutters[i] = ref;
+    width = width || this.state.gutterWidth;
 
     let range = dates.range(start, end, 'day')
 
@@ -77,7 +100,7 @@ let TimeGrid = React.createClass({
           , eEnd = get(event, endAccessor);
 
         if (
-             get(event, allDayAccessor)
+          get(event, allDayAccessor)
           || !dates.eq(eStart, eEnd, 'day')
           || (dates.isJustDate(eStart) && dates.isJustDate(eEnd)))
         {
@@ -90,45 +113,35 @@ let TimeGrid = React.createClass({
 
     allDayEvents.sort((a, b) => sortEvents(a, b, this.props))
 
-    let segments = allDayEvents.map(evt => eventSegments(evt, start, end, this.props))
-    let { levels } = eventLevels(segments)
+    let {first, last} = endOfRange(range);
+
+    let segments = allDayEvents.map(evt => eventSegments(evt, first, last, this.props))
+
+    let gutterRef = ref => this._gutters[1] = ref && findDOMNode(ref);
 
     return (
       <div className='rbc-time-view'>
-        <div ref='headerCell' className='rbc-time-header'>
-          <div className='rbc-row'>
-            <div ref={addGutterRef(0)} className='rbc-gutter-cell'/>
-            { this.renderHeader(range) }
-          </div>
-          <div className='rbc-row'>
-            <div ref={addGutterRef(1)} className='rbc-gutter-cell'>
-              { message(messages).allDay }
-            </div>
-            <div ref='allDay' className='rbc-allday-cell'>
-              <BackgroundCells
-                slots={range.length}
-                container={()=> this.refs.allDay}
-                selectable={this.props.selectable}
-              />
-              <div style={{ zIndex: 1, position: 'relative' }}>
-                { this.renderAllDayEvents(range, levels) }
-              </div>
-            </div>
-          </div>
-        </div>
+        {
+          this.renderHeader(range, segments, width)
+        }
         <div ref='content' className='rbc-time-content'>
-          <TimeGutter ref='gutter' {...this.props}/>
+          <TimeColumn
+            {...this.props}
+            showLabels
+            style={{ width }}
+            ref={gutterRef}
+            className='rbc-time-gutter'
+          />
           {
-            this.renderEvents(range, rangeEvents)
+            this.renderEvents(range, rangeEvents, this.props.now)
           }
         </div>
       </div>
     );
-  },
+  }
 
-  renderEvents(range, events){
-    let { endAccessor, startAccessor, components } = this.props;
-    let today = new Date();
+  renderEvents(range, events, today){
+    let { min, max, endAccessor, startAccessor, components } = this.props;
 
     return range.map((date, idx) => {
       let daysEvents = events.filter(
@@ -138,8 +151,10 @@ let TimeGrid = React.createClass({
       )
 
       return (
-        <DaySlot
+        <DayColumn
           {...this.props }
+          min={dates.merge(date, min)}
+          max={dates.merge(date, max)}
           eventComponent={components.event}
           className={cn({ 'rbc-now': dates.eq(date, today, 'day') })}
           style={segStyle(1, this._slots)}
@@ -149,11 +164,10 @@ let TimeGrid = React.createClass({
         />
       )
     })
-  },
+  }
 
   renderAllDayEvents(range, levels){
-    let first = range[0]
-      , last = range[range.length - 1];
+    let { first, last } = endOfRange(range);
 
     while (levels.length < MIN_ROWS )
       levels.push([])
@@ -165,6 +179,7 @@ let TimeGrid = React.createClass({
         startAccessor={this.props.startAccessor}
         endAccessor={this.props.endAccessor}
         allDayAccessor={this.props.allDayAccessor}
+        eventPropGetter={this.props.eventPropGetter}
         onSelect={this._selectEvent}
         slots={this._slots}
         key={idx}
@@ -173,13 +188,63 @@ let TimeGrid = React.createClass({
         end={last}
       />
     )
-  },
+  }
 
-  renderHeader(range){
+  renderHeader(range, segments, width) {
+    let { messages, rtl } = this.props;
+    let { isOverflowing } = this.state || {};
+
+    let { levels } = eventLevels(segments);
+    let style = {};
+
+    if (isOverflowing)
+      style[rtl ? 'marginLeft' : 'marginRight'] = scrollbarSize() + 'px';
+
+    return (
+      <div
+        ref='headerCell'
+        className={cn(
+          'rbc-time-header',
+          isOverflowing && 'rbc-overflowing'
+        )}
+        style={style}
+      >
+        <div className='rbc-row'>
+          <div
+            className='rbc-label rbc-header-gutter'
+            style={{ width }}
+          />
+          { this.renderHeaderCells(range) }
+        </div>
+        <div className='rbc-row'>
+          <div
+            ref={ref => this._gutters[0] = ref}
+            className='rbc-label rbc-header-gutter'
+            style={{ width }}
+          >
+            { message(messages).allDay }
+          </div>
+          <div ref='allDay' className='rbc-allday-cell'>
+            <BackgroundCells
+              slots={range.length}
+              container={()=> this.refs.allDay}
+              selectable={this.props.selectable}
+            />
+            <div style={{ zIndex: 1, position: 'relative' }}>
+              { this.renderAllDayEvents(range, levels) }
+            </div>
+          </div>
+        </div>
+      </div>
+    )
+  }
+
+  renderHeaderCells(range){
     let { dayFormat, culture } = this.props;
 
     return range.map((date, i) =>
-      <div key={i}
+      <div
+        key={i}
         className='rbc-header'
         style={segStyle(1, this._slots)}
       >
@@ -188,40 +253,41 @@ let TimeGrid = React.createClass({
         </a>
       </div>
     )
-  },
+  }
 
   _headerClick(date, e){
     e.preventDefault()
     notify(this.props.onNavigate, [navigate.DATE, date])
-  },
+  }
 
   _selectEvent(...args){
     notify(this.props.onSelectEvent, args)
-  },
+  }
 
-  _adjustGutter() {
-    let isRtl = this.props.rtl;
-    let header = this.refs.headerCell;
-    let width = this._gutterWidth
-    let isOverflowing = this.refs.content.scrollHeight > this.refs.content.clientHeight;
+  measureGutter() {
+    let width = this.state.gutterWidth;
+    let gutterCells = this._gutters;
 
-    this._gutterWidth = getWidth(findDOMNode(this.refs.gutter));
+    if (!width) {
+      width = Math.max(...gutterCells.map(getWidth));
 
-    if (width !== this._gutterWidth) {
-      width = this._gutterWidth + 'px';
-      this._gutters.forEach(node => node.style.width = width)
-    }
-
-    if (isOverflowing) {
-      classes.addClass(header, 'rbc-header-overflowing')
-      this.refs.headerCell.style[!isRtl ? 'marginLeft' : 'marginRight'] = '';
-      this.refs.headerCell.style[isRtl ? 'marginLeft' : 'marginRight'] = scrollbarSize() + 'px';
-    } else {
-      classes.removeClass(header, 'rbc-header-overflowing')
+      if (width) {
+        this.setState({ gutterWidth: width })
+      }
     }
   }
 
-});
+  checkOverflow() {
+    if (this._updatingOverflow) return;
 
+    let isOverflowing = this.refs.content.scrollHeight > this.refs.content.clientHeight;
 
-export default TimeGrid
+    if (this.setState.isOverflowing !== isOverflowing) {
+      this._updatingOverflow = true;
+      this.setState({ isOverflowing }, () => {
+        this._updatingOverflow = false;
+      })
+    }
+  }
+
+}