+++ src/utils/dates.js
@@ -121,9 +121,12 @@ let dates = Object.assign(dateMath, {
     if (!unit)
       return Math.abs(+dateA - +dateB)
 
-    return Math.abs(
+    // the .round() handles an edge case
+    // with DST where the total won't be exact
+    // since one day in the range may be shorter/longer by an hour
+    return Math.round(Math.abs(
       (+dates.startOf(dateA, unit) / MILLI[unit]) - (+dates.startOf(dateB, unit) / MILLI[unit])
-    )
+    ))
   },
 
   total(date, unit) {